@@ -1,6 +1,7 @@
 plugins {
-    id("com.github.johnrengelman.shadow") version "6.1.0"
-    id("io.micronaut.application") version "1.4.2"
+    id("com.github.johnrengelman.shadow") version "7.1.2"
+    id("io.micronaut.application") version "3.6.1"
+    id("io.micronaut.test-resources") version "3.6.1"
 }
 
 version = "0.1"
@@ -17,10 +18,14 @@ micronaut {
         incremental(true)
         annotations("example.*")
     }
+    testResources {
+        clientTimeout = 360
+    }
 }
 
 dependencies {
     annotationProcessor("io.micronaut.data:micronaut-data-processor")
+    implementation(platform("com.oracle.database.jdbc:ojdbc-bom:21.7.0.0"))
     implementation("io.micronaut.data:micronaut-data-jdbc")
     implementation("io.micronaut:micronaut-validation")
     implementation("io.micronaut:micronaut-runtime")
@@ -28,10 +33,9 @@ dependencies {
     runtimeOnly("ch.qos.logback:logback-classic")
     runtimeOnly("io.micronaut.sql:micronaut-jdbc-hikari")
     runtimeOnly("com.oracle.database.jdbc:ojdbc8")
-    runtimeOnly("com.oracle.database.security:oraclepki:21.1.0.0")
-    runtimeOnly("com.oracle.database.security:osdt_cert:21.1.0.0")
-    runtimeOnly("com.oracle.database.security:osdt_core:21.1.0.0")
-    compileOnly("org.graalvm.nativeimage:svm")
+    runtimeOnly("com.oracle.database.security:oraclepki")
+    runtimeOnly("com.oracle.database.security:osdt_cert")
+    runtimeOnly("com.oracle.database.security:osdt_core")
     runtimeOnly("io.micronaut.flyway:micronaut-flyway")
 }
 
@@ -40,7 +44,8 @@ application {
 }
 
 java {
-    sourceCompatibility = JavaVersion.toVersion("11")
-    targetCompatibility = JavaVersion.toVersion("11")
+    sourceCompatibility = JavaVersion.toVersion("17")
+    targetCompatibility = JavaVersion.toVersion("17")
 }
+
 
