@@ -1,11 +1,11 @@
 package example.atp;
 
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
 import io.micronaut.runtime.EmbeddedApplication;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Assertions;
-
-import javax.inject.Inject;
+import jakarta.inject.Inject;
 
 @MicronautTest
 public class ExampleAtpTest {
