@@ -3,14 +3,5 @@ package example.atp.domain;
 import io.micronaut.core.annotation.Introspected;
 
 @Introspected
-public class NameDTO {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
+public record NameDTO(String name) {
 }
