@@ -1,41 +1,17 @@
 package example.atp.domain;
 
-import io.micronaut.core.annotation.Creator;
+import io.micronaut.core.annotation.Nullable;
 import io.micronaut.data.annotation.GeneratedValue;
+import io.micronaut.data.annotation.GeneratedValue.Type;
 import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.MappedEntity;
 
 @MappedEntity
-public class Owner {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-    private String name;
-    private int age;
-
-    @Creator
-    public Owner(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
+public record Owner(
+    @Id @GeneratedValue(Type.IDENTITY) @Nullable Long id, 
+    String name, 
+    int age) {
+    public Owner(String name, int age) {
+        this(null, name, age);
     }
 }
