@@ -1,52 +1,35 @@
 package example.atp.domain;
 
-import io.micronaut.core.annotation.Creator;
+import java.util.UUID;
+
+import example.atp.domain.Owner;
+import io.micronaut.core.annotation.Nullable;
 import io.micronaut.data.annotation.AutoPopulated;
 import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.MappedEntity;
 import io.micronaut.data.annotation.Relation;
-import io.micronaut.core.annotation.Nullable;
-import java.util.UUID;
 
 @MappedEntity
-public class Pet {
-
-    @Id
-    @AutoPopulated
-    private UUID id;
-    private String name;
+public record Pet(
+    @Id @AutoPopulated @Nullable UUID id, 
+    String name, 
     @Relation(Relation.Kind.MANY_TO_ONE)
-    private Owner owner;
-    private PetType type = PetType.DOG;
-
-    @Creator
-    public Pet(String name, @Nullable Owner owner) {
-        this.name = name;
-        this.owner = owner;
+    Owner owner, 
+    @Nullable
+    PetType type) {
+        
+    public Pet {
+        if (type == null) {
+            type = PetType.DOG;
+        }
     }
 
-    public Owner getOwner() {
-        return owner;
+    public Pet(String name, Owner owner) {
+        this(null, name, owner, null);
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public PetType getType() {
-        return type;
-    }
-
-    public void setType(PetType type) {
-        this.type = type;
+    public Pet(String name, Owner owner, PetType type) {
+        this(null, name, owner, type);
     }
 
     public enum PetType {
